+ fix delete coin from favourites

+++ app/api/services/db_services.py
@@ -183,10 +183,13 @@ def delete_favourite_coin(current_user: User, db: Session, coin: Cryptocurrency)
         .filter(UserFavouriteCrypto.coin_id == coin.id)
         .first()
     )
-    db.delete(user_with_fav_crypto)
-    db.commit()
-    current_favourite_cryptos = user_favourite_cryptocurrency(current_user, db)
-    return current_favourite_cryptos
+    if bool(user_with_fav_crypto):
+        db.delete(user_with_fav_crypto)
+        db.commit()
+        current_favourite_cryptos = user_favourite_cryptocurrency(current_user, db)
+        return current_favourite_cryptos
+    else:
+        return {"msg": "You dont have this coin."}
 
 
 def update_favourites_coins(current_user: User, coin: Cryptocurrency, db: Session):
